More WIP.

+++ app/features/EventEdit/features/content/components/ContentTypeContainer.js
@@ -26,32 +26,42 @@ class ContentTypeContainer extends Component {
 	}
 
 	handleChange(changeEvent) {
-		if (typeof changeEvent.target === 'object') {
-			const { contentIndex, handleChange } = this.props
-			const { name, value } = changeEvent.target
-			handleChange(contentIndex, name, value)
-		}
+		const { event } = this.props
 
-		this.autoSave()
+		if (event.rules.content.canEdit || event.rules.content.canAdd) {
+			if (typeof changeEvent.target === 'object') {
+				const { contentIndex, handleChange } = this.props
+				const { name, value } = changeEvent.target
+				handleChange(contentIndex, name, value)
+			}
+
+			this.autoSave()
+		}
 	}
 
 	handleRemove(removeEvent) {
 		const { contentIndex, content, handleRemove, event } = this.props
-		handleRemove(contentIndex, event.item.id, content.content_id)
+
+		if (event.rules.content.canDelete) {
+			handleRemove(contentIndex, event.item.id, content.content_id)
+		}
 	}
 
 	autoSave() {
-		const { content, handleUpdate } = this.props
+		const { event, content, handleUpdate } = this.props
 
 		if (typeof this.cancelSave === 'number') {
 			clearTimeout(this.cancelSave)
 			this.cancelSave = null
 		}
-		this.cancelSave = setTimeout(::this.handleUpdateClick, AUTO_SAVE_TIMEOUT)
+
+		if (event.rules.content.canEdit || event.rules.content.canAdd) {
+			this.cancelSave = setTimeout(::this.handleUpdateClick, AUTO_SAVE_TIMEOUT)
+		}
 	}
 
 	render() {
-		const { dispatch, contentIndex, content, references, plans } = this.props
+		const { event, dispatch, contentIndex, content, references, plans } = this.props
 
 		let InnerContainer = null
 		switch (content.type) {
@@ -104,7 +114,7 @@ class ContentTypeContainer extends Component {
 			<div className='content-type-text'>
 				<Row>
 					<div className='medium-12'>
-						{content.type.toUpperCase()} <a className='right' onClick={::this.handleRemove}><img src={`/images/${RevManifest['thin-x.png']}`} /></a>
+						{content.type.toUpperCase()} <a disabled={!event.rules.content.canDelete} className='right' onClick={::this.handleRemove}><img src={`/images/${RevManifest['thin-x.png']}`} /></a>
 						<div className='form-body'>
 							{InnerContainer}
 							{ content.isDirty ? 'D' : '.'}
